% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rfVIP_master.R
\name{rfeTrainingSplit}
\alias{rfeTrainingSplit}
\title{rfeTrainingSplit}
\usage{
rfeTrainingSplit(x, y, nSeed = 3456)
}
\arguments{
\item{x}{dataframe containing x values}

\item{y}{vector containing y values}

\item{nSeed}{integer containing seed value, defaults to 3456}
}
\value{
a list containing values in the following order: training y, training x,
test-train index, full y, full x
}
\description{
This function does the training-test split of data
}
\examples{
splitList <- rfeTrainingSplit(x,y,nSeed=3456)
}

