% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rfVIP_master.R
\name{rfeTuningFunction}
\alias{rfeTuningFunction}
\title{rfeTuningFunction}
\usage{
rfeTuningFunction(train.x, train.y, nSeed = 3456, nCores = 1)
}
\arguments{
\item{train.x}{dataframe containing x values}

\item{train.y}{vector containing y values}

\item{nSeed}{integer containing seed value, defaults to 3456}

\item{nCores}{integer designating the number of cores for parallel processing, defaults to 1}
}
\value{
an integer containing the mtry value with least error
}
\description{
This function does one round of random forest modeling and finds the best performing mtry
}
\examples{
m.try <- rfeTuningFunction(train.x,train.y,nSeed=3456,nCores=4)
}

